+++ tests/test_trainers_args.py
@@ -31,7 +31,6 @@ from trl import (
     SFTConfig,
     SFTTrainer,
 )
-from trl.experimental.nash_md import NashMDConfig, NashMDTrainer
 
 from .testing_utils import TrlTestCase
 
@@ -148,28 +147,6 @@ class TestTrainerArg(TrlTestCase):
         assert trainer.args.model_init_kwargs == {"trust_remote_code": True}
         assert trainer.args.ref_model_init_kwargs == {"trust_remote_code": True}
         assert trainer.args.dataset_num_proc == 4
-
-    @pytest.mark.parametrize("mixtures_coef_list", [False, True])
-    def test_nash_md(self, mixtures_coef_list):
-        model_id = "trl-internal-testing/tiny-Qwen2ForCausalLM-2.5"
-        tokenizer = AutoTokenizer.from_pretrained(model_id)
-        model = AutoModelForCausalLM.from_pretrained(model_id)
-        ref_model = AutoModelForCausalLM.from_pretrained(model_id)
-        reward_model = AutoModelForSequenceClassification.from_pretrained(model_id, num_labels=1)
-        dataset = load_dataset("trl-internal-testing/zen", "standard_prompt_only", split="train")
-        training_args = NashMDConfig(
-            self.tmp_dir,
-            mixture_coef=0.5 if not mixtures_coef_list else [0.5, 0.6],
-        )
-        trainer = NashMDTrainer(
-            args=training_args,
-            processing_class=tokenizer,
-            model=model,
-            ref_model=ref_model,
-            reward_funcs=reward_model,
-            train_dataset=dataset,
-        )
-        assert trainer.args.mixture_coef == (0.5 if not mixtures_coef_list else [0.5, 0.6])
 
     @pytest.mark.parametrize("beta_list", [False, True])
     def test_online_dpo(self, beta_list):
