+++ trl/trainer/nash_md_config.py
@@ -15,34 +15,15 @@
 import warnings
 from dataclasses import dataclass
 
-from ..experimental.nash_md import NashMDConfig as ExperimentalNashMDConfig
+from ..experimental.nash_md import NashMDConfig as _NashMDConfig
 
 
 @dataclass
-class NashMDConfig(ExperimentalNashMDConfig):
-    r"""
-    Configuration class for the [`NashMDTrainer`].
-
-    <Deprecated version="0.25.0">
-
-    This class has been moved to `trl.experimental.nash_md.NashMDConfig` and will be removed in TRL 0.29.0.
-    Please update your imports:
-    ```python
-    from trl.experimental.nash_md import NashMDConfig
-    ```
-
-    For more details, see: https://github.com/huggingface/trl/issues/4223
-
-    </Deprecated>
-    """
-
+class NashMDConfig(_NashMDConfig):
     def __post_init__(self):
         warnings.warn(
-            "NashMDConfig has been moved to trl.experimental.nash_md.NashMDConfig and will be removed from "
-            "trl.trainer in TRL 0.29.0. Please update your imports to: "
-            "`from trl.experimental.nash_md import NashMDConfig`. "
-            "For more details, see: https://github.com/huggingface/trl/issues/4223",
-            FutureWarning,
-            stacklevel=2,
+            "The `NashMDConfig` is now located in `trl.experimental`. Please update your imports to "
+            "`from trl.experimental.nash_md import NashMDConfig`. The current import path will be removed and no "
+            "longer supported in TRL 0.29. For more information, see https://github.com/huggingface/trl/issues/4223."
         )
         super().__post_init__()
